requirements.txt
# django
django
djangorestframework
django-environ
psycopg2-binary
djangorestframework-simplejwt
django_ratelimit
django-celery-beat
requests
django-storages
Pillow
django-cors-headers

# Celery redis
boto3
celery
redis

# env
python-dotenv

# gunicron tools
gunicorn
uvicorn
uvloop
httptools